Move `copyfile` to fs.py
