+++ dvc/remote/local.py
@@ -21,7 +21,7 @@ from dvc.remote.base import STATUS_NEW
 from dvc.scheme import Schemes
 from dvc.scm.tree import is_working_tree
 from dvc.system import System
-from dvc.utils import copyfile
+from dvc.utils.fs import copyfile
 from dvc.utils import file_md5
 from dvc.utils import relpath
 from dvc.utils import tmp_fname
